+++ tests/integration/test_meteor_interface_integration.py
@@ -1,5 +1,8 @@
 import os
 
+import numpy as np
+import xarray as xr
+
 from meteor.ensemble_output import EnsembleOutput, VariableOutput
 from meteor.meteor_interface import MeteorInterface
 
@@ -13,7 +16,21 @@ def test_meteor_interface_integration_pr(test_data_dir):
     meteor = MeteorInterface("CanESM5", variables=["pr"], cache_dir=cache_path)
     assert meteor._is_trained["pr"] is False
     # Train the model using the simple dataset
-    meteor.train(variable_configs={"pr": {"n_modes_noise": 4, "lag_order": 1}})
+    meteor.train(
+        variable_configs={
+            "pr": {
+                "n_modes_noise": 4,
+                "lag_order": 1,
+                "n_modes_pattern": 3,
+                "use_exog": "none",
+                "use_picontrol_baseline": True,
+                "transform": True,
+                "transform_type": "gamma",
+            }
+        },
+        auto=False,
+        verbose=False,
+    )
     assert meteor._is_trained["pr"] is True
 
     # Generate ensemble output
@@ -57,6 +74,34 @@ def test_meteor_interface_integration_pr(test_data_dir):
     # assert ensemble.pr.sizes["lat"] == 5
     # assert ensemble.pr.sizes["ensemble"] == 10
     # assert ensemble.pr.sizes["time"] == 102  # 2000 to 2100 inclusive
+    temp_ts_mock = xr.DataArray(
+        data=np.linspace(0, 4, 351),  # Mock temperature time series
+        dims=["year"],
+        coords={"year": np.arange(1750, 2101)},  # Years from 1750 to 2100
+    )
+
+    ensemble_scaled = meteor.generate_ensemble_outputs(
+        scenario="ssp245",
+        start_year=2015,
+        end_year=2035,
+        n_realizations=2,
+        timeseries=["global", "regional:EAS"],
+        gridded={"annual": list(range(2020, 2026))},
+        temp_scaling_ts=temp_ts_mock,
+    )
+    assert isinstance(ensemble_scaled, EnsembleOutput)
+    assert ensemble_scaled.variables is not None
+    assert ensemble_scaled.metadata["scenario"] == "ssp245"
+    assert ensemble_scaled.metadata["n_realizations"] == 2
+    assert ensemble_scaled.metadata["model"] == "CanESM5"
+    assert ensemble_scaled.metadata["year_range"] == "2015-2035"
+    assert isinstance(ensemble_scaled.variables["pr"], VariableOutput)
+    base_value = ensemble["pr"].timeseries["global"][0][0]
+    assert not np.allclose(
+        ensemble["pr"].timeseries["global"][0] - base_value,
+        ensemble_scaled["pr"].timeseries["global"][0] - base_value,
+    )
+    # assert global_mean()
 
 
 def test_meteor_interface_integration_tas(test_data_dir):
@@ -99,14 +144,14 @@ def test_meteor_interface_integration_tas(test_data_dir):
     # Check that the ensemble output has the expected shape
     assert isinstance(ensemble, EnsembleOutput)
     assert ensemble.variables is not None
-    print(ensemble.variables)
-    print(ensemble.metadata)
+    # print(ensemble.variables)
+    # print(ensemble.metadata)
     assert ensemble.metadata["scenario"] == "ssp245"
     assert ensemble.metadata["n_realizations"] == 10
     assert ensemble.metadata["model"] == "CanESM5"
     assert ensemble.metadata["year_range"] == "2015-2035"
     assert isinstance(ensemble.variables["tas"], VariableOutput)
-    print(ensemble.variables["tas"].impacts)
+    # print(ensemble.variables["tas"].impacts)
     assert ensemble.variables["tas"].impacts["hdd"]["global"].shape == (
         10,
         21,
@@ -114,3 +159,36 @@ def test_meteor_interface_integration_tas(test_data_dir):
     # assert ensemble.tas.sizes["lat"] == 5
     # assert ensemble.tas.sizes["ensemble"] == 10
     # assert ensemble.tas.sizes["time"] == 102  # 2000 to 2100 inclusive
+    temp_ts_mock = xr.DataArray(
+        data=np.linspace(0, 4, 351),  # Mock temperature time series
+        dims=["year"],
+        coords={"year": np.arange(1750, 2101)},  # Years from 1750 to 2100
+    )
+
+    ensemble_scaled = meteor.generate_ensemble_outputs(
+        scenario="ssp245",
+        start_year=2015,
+        end_year=2035,
+        n_realizations=2,
+        timeseries=["global", "regional:EAS"],
+        gridded={"annual": list(range(2020, 2026))},
+        temp_scaling_ts=temp_ts_mock,
+    )
+    assert isinstance(ensemble_scaled, EnsembleOutput)
+    assert ensemble_scaled.variables is not None
+    assert ensemble_scaled.metadata["scenario"] == "ssp245"
+    assert ensemble_scaled.metadata["n_realizations"] == 2
+    assert ensemble_scaled.metadata["model"] == "CanESM5"
+    assert ensemble_scaled.metadata["year_range"] == "2015-2035"
+    assert isinstance(ensemble_scaled.variables["tas"], VariableOutput)
+
+    # Testing that scaling mostly worked
+    pred_scaled = temp_ts_mock.values[2015 - 1750 : 2036 - 1750]
+    ann_mean_scaled = np.average(
+        ensemble_scaled["tas"].timeseries["global"][0].reshape(-1, 12), axis=1
+    )
+    ann_mean_unscaled = np.average(
+        ensemble["tas"].timeseries["global"][0].reshape(-1, 12), axis=1
+    )
+    assert np.allclose(pred_scaled, ann_mean_scaled, atol=0.5)
+    assert not np.allclose(ann_mean_scaled, ann_mean_unscaled, atol=1.2)